Move to allocator submodule
